squash! [bugfix] remove observer on 'selected' update
An Ember PromiseProxy hasOwnProperty 'content' only when the
underlying promise is fulfilled.

Therefore, we only attach an observer on the `content` property when a
PromiseProxy becomes fulfilled. This also guards for when the
component is already destroyed, since we do not want to attach an
observer after destrction.

In tests we cleaned up PromiseProxy instantiations:

- Removes initializing the `content` property of the PromiseProxy
- Removes setting the `promise` property of the PromiseProxy on update

Once the underlying promise of a PromiseProxy is resolved, we only
need to update the `content` property to trigger the observers on the
`.content` property.

We updated the general PromiseProxy test to exercise a promise moving
from pending to fulfilled state. We also test updating the proxy content to a
present value as well as updating the proxy content to a blank value.